requirements.txt
flask
imageio
scipy
numpy
opencv-python
python-telegram-bot
